First commit.
OK so this a little program I wrote to parse the Brussels Airlines page (https://www.brusselsairlines.com/nl-nl/bestemmingen/Default.aspx) with all their flights to a list. 
Then i pick a random country from it and open it in a browser, that's our vacation destination. Easy as py.